Setup Docker infrastructure for v1.13 [more]
Our production servers are Red Hat Enterprise
Linux 7.5 and only support an older version
of docker, docker 1.13. Note that docker-compose
is the most up-to-date (docker-compose 1.22),
so the docker version is the bottleneck.
Docker 1.13 does not have multi-stage
build, so we recreated that manually.

We perform a multi-stage(image) build in order
to position ourselves for sharing images (even
if only internally) in the future (docker
registry for instance). This build pattern:
- prevents sensitive information from leaking
  into the main image
- sets ourselves up to reduce the size of the final image

Closes #81