Assimilate cd2h-datamodel [more]
Keeping the datamodel and the invenio instance separate
was an unnecessary pain: branches had to be kept in sync
and tests were written in a different repo then the
tested code. Everything depends on the datamodel
and our flavour of Invenio does not prioritize datamodel
swapping i.e. there is no 'modularity' argument in our
case.

This assimilation also:
- Consolidates configuration in the main config.py
- Fixes isort import bug
- Consolidates fixtures