[#367] Pick up failing builds in Jenkins [details]
There is a bug in docker-compose where the exit code of the service
is not returned and exit 0 is always returned when containers are downed.
We now use scripts/jenkins_ci.sh to run the continuous
integration tests and overcome this bug by looking at the exit code in the
output.